Report the chosen mode as LineNoise.mode.
Change the way that LineNoise is initialized, so that we decide
immediately what mode we need to use (i.e. which path in getLine
we are going to take).  Add an enum LineNoise.Mode with
unsupportedTTY / supportedTTY / notATTY as options.  Add
LineNoise.mode to report the value that we chose.

This has two advantages: 1) we don't check isatty and
ProcessInfo.processInfo.environment["TERM"] with every call to getLine,
and 2) client code can use the mode if it wants to behave differently
for each case.

Merge the two initializers at the same time (one was just setting
default args for the other).