Add an option to preserve user edits to history.
This adds a new flag, LineNoise.preserveHistoryEdits.  The default is
to follow existing behavior, which is that any edits by the user to a
line in the history will be discarded if the user moves forward or back
in the history without pressing Enter.  Setting this flag to true, will
cause all history edits will be preserved, even if the user is just
moving backwards and forwards in the history.  This matches bash's
(presumably readline's) behavior.